+++ bot/main.py
@@ -41,6 +41,7 @@ nonebot.load_plugin("plugins.friend")
 nonebot.load_plugin("plugins.member")
 nonebot.load_plugin("plugins.mute")
 nonebot.load_plugin("plugins.shortcuts")
+nonebot.load_plugin("plugins.auto_recall")
 # TODO: 后续改为 load_plugins("plugins") 自动扫描所有插件
 
 # --- 启动时打印已加载插件 ---